refactor QCOM emulator structure
