+++ test/mockgpu/qcom/uop_runner.py
@@ -1,14 +1,5 @@
-"""Optional native-CPU execution for decoded, straight-line IR3 ALU blocks.
-
-This is deliberately a translator for :class:`IR3Instruction`, not a fallback
-to the source Tensor/UOp graph.  Unsupported machine instructions return
-``None`` from :meth:`IR3UOpRunner.try_run`, leaving the IR3 interpreter as the
-single correctness authority.  Each lowered opcode mirrors the exact branch
-the interpreter takes for it, including which branches ignore encoded source
-modifiers (``_IGNORE_MODS``) and which apply them.
-"""
+"""Lower decoded straight-line IR3 ALU blocks to CPU UOps."""
 import array
-from collections import OrderedDict
 from dataclasses import dataclass
 from typing import Any, TypeGuard
 
@@ -44,37 +35,6 @@ _MAD_INT = {'mad.u16', 'madsh.u16', 'mad.s16', 'mad.u24', 'mad.s24'}
 
 class UnsupportedIR3Block(Exception):
   """The decoded block needs the ordinary machine-code interpreter."""
-
-
-class IR3UOpLoopTimeout(RuntimeError):
-  """Internal signal: scalar-replay this staged loop through its exact fuel limit."""
-  def __init__(self, start_pc: int):
-    super().__init__(f'IR3 native loop exhausted fuel at PC {start_pc}')
-    self.start_pc = start_pc
-
-
-@dataclass
-class IR3UOpRunnerStats:
-  """Small, resettable counters for the optional native block and loop paths."""
-  attempts: int = 0
-  runs: int = 0
-  native_calls: int = 0
-  fallbacks: int = 0
-  cache_hits: int = 0
-  compiled: int = 0
-  iterations: int = 0
-  load_checks: int = 0
-  load_rejections: int = 0
-  block_attempts: int = 0
-  block_runs: int = 0
-  block_compiles: int = 0
-  block_declines: int = 0
-  cache_evictions: int = 0
-
-  def reset(self) -> None:
-    self.attempts = self.runs = self.native_calls = self.fallbacks = self.cache_hits = 0
-    self.compiled = self.iterations = self.load_checks = self.load_rejections = 0
-    self.block_attempts = self.block_runs = self.block_compiles = self.block_declines = self.cache_evictions = 0
 
 
 def _register(value: Any) -> TypeGuard[Register]:
@@ -423,200 +383,3 @@ class _Lowerer:
     return IR3UOpBlock(len(self.instructions), self.lanes, self.keys,
                        tuple((reg, self.slot[reg]) for reg in sorted(self.dirty)), runtime, regfile,
                        regfile.as_memoryview(force_zero_copy=True).cast('I'))
-
-
-class IR3UOpRunner:
-  """Compile supported decoded IR3 ALU blocks and conservative single-lane natural loops to CPU UOps."""
-  def __init__(self, min_instructions: int = 8, max_instructions: int = 64, max_register_slots: int = 128,
-               max_average_register_slots: int = 40, max_compiled_blocks: int = 512, max_block_locations: int = 8192,
-               max_programs: int = 128, max_compiled_loops: int = 32, max_loop_locations: int = 64,
-               max_narrow_compiled_blocks: int | None = None, max_regular_narrow_compiled_blocks: int | None = None):
-    if not 1 <= min_instructions <= max_instructions: raise ValueError('invalid native IR3 block instruction limits')
-    if not 1 <= max_average_register_slots <= max_register_slots: raise ValueError('invalid native IR3 block register limits')
-    if min(max_compiled_blocks, max_block_locations, max_programs, max_compiled_loops, max_loop_locations) < 1:
-      raise ValueError('invalid native IR3 cache limits')
-    if max_narrow_compiled_blocks is None: max_narrow_compiled_blocks = min(384, max_compiled_blocks)
-    if not 1 <= max_narrow_compiled_blocks <= max_compiled_blocks: raise ValueError('invalid narrow IR3 compile budget')
-    if max_regular_narrow_compiled_blocks is None: max_regular_narrow_compiled_blocks = min(128, max_narrow_compiled_blocks)
-    if not 1 <= max_regular_narrow_compiled_blocks <= max_narrow_compiled_blocks:
-      raise ValueError('invalid regular narrow IR3 compile budget')
-    self.min_instructions, self.max_instructions = min_instructions, max_instructions
-    self.max_register_slots, self.max_average_register_slots = max_register_slots, max_average_register_slots
-    self.max_compiled_blocks, self.max_block_locations, self.max_programs = max_compiled_blocks, max_block_locations, max_programs
-    self.max_compiled_loops, self.max_loop_locations = max_compiled_loops, max_loop_locations
-    self.max_narrow_compiled_blocks = max_narrow_compiled_blocks
-    self.max_regular_narrow_compiled_blocks = max_regular_narrow_compiled_blocks
-    self.cache: OrderedDict[tuple[int, int, int, int], IR3UOpBlock] = OrderedDict()
-    self.uncompilable: OrderedDict[tuple[int, int, int, int], None] = OrderedDict()
-    self.program_blocks: OrderedDict[int, tuple[tuple[IR3Instruction, ...], dict[int, int]]] = OrderedDict()
-    self.program_policy: OrderedDict[tuple[int, bool], tuple[tuple[IR3Instruction, ...], bool]] = OrderedDict()
-    # Structurally identical blocks (same register layout, lanes, and instruction forms) share one
-    # compiled program, so equivalent sequences in different shaders never recompile.
-    self.compiled: OrderedDict[tuple, IR3UOpBlock] = OrderedDict()
-    self.compiled_classes: dict[tuple, str] = {}
-    self.loop_cache: OrderedDict[tuple[int, int], tuple[tuple[IR3Instruction, ...], Any]] = OrderedDict()
-    self.loop_uncompilable: OrderedDict[tuple[int, int], None] = OrderedDict()
-    self.compiled_loops: OrderedDict[tuple, Any] = OrderedDict()
-    self.stats = IR3UOpRunnerStats()
-    self._vmem: Buffer | None = None
-
-  @staticmethod
-  def _put_lru(cache: OrderedDict, key, value, limit: int):
-    cache[key] = value
-    cache.move_to_end(key)
-    return cache.popitem(last=False) if len(cache) > limit else None
-
-  @staticmethod
-  def _get_lru(cache: OrderedDict, key):
-    if key not in cache: return None
-    cache.move_to_end(key)
-    return cache[key]
-
-  def _evict_compiled_class(self, block_class: str) -> bool:
-    for signature in tuple(self.compiled):
-      if self.compiled_classes.get(signature) != block_class: continue
-      block = self.compiled.pop(signature)
-      del self.compiled_classes[signature]
-      for location, cached_block in tuple(self.cache.items()):
-        if cached_block is block: del self.cache[location]
-      self.stats.cache_evictions += 1
-      return True
-    return False
-
-  @staticmethod
-  def _supported(inst: IR3Instruction) -> bool:
-    if inst.name == 'nop': return True
-    if inst.name not in _NATIVE or not _register(inst.dst) or inst.sat or (inst.name == 'mov' and inst.rounding): return False
-    return not any(isinstance(src, tuple) and src[0] in {'rel', 'relr', 'relhr'} for src in inst.srcs)
-
-  @classmethod
-  def _loop_shape(cls, program: tuple[IR3Instruction, ...], start_pc: int) -> Any:
-    from test.mockgpu.qcom.loop_runner import loop_shape
-    return loop_shape(program, start_pc, cls._supported)
-
-  @classmethod
-  def has_loop(cls, program: tuple[IR3Instruction, ...]) -> bool:
-    from test.mockgpu.qcom.loop_runner import has_loop
-    return has_loop(program, cls._supported)
-
-  @staticmethod
-  def _select_memory_bounds(regs: dict[Register, list[int]], bounds: tuple[tuple[int, int], ...]) -> tuple[tuple[int, int], ...] | None:
-    from test.mockgpu.qcom.loop_runner import _select_memory_bounds
-    return _select_memory_bounds(regs, bounds)
-
-  def _vmem_buffer(self) -> Buffer:
-    from test.mockgpu.qcom.loop_runner import _vmem_buffer
-    return _vmem_buffer(self)
-
-  def _loop_block(self, program: tuple[IR3Instruction, ...], start_pc: int) -> Any:
-    from test.mockgpu.qcom.loop_runner import loop_block
-    return loop_block(self, program, start_pc)
-
-  def try_run_loop(self, program: tuple[IR3Instruction, ...], start_pc: int, regs: dict[Register, list[int]], exec_mask: list[bool], *,
-                   check_range=None, memory_bounds: tuple[tuple[int, int], ...] | None = None,
-                   max_steps: int | None = None) -> tuple[int, int] | None:
-    """Run one eligible decoded-IR3 natural loop in a single native CPU call."""
-    from test.mockgpu.qcom.loop_runner import try_run_loop
-    return try_run_loop(self, program, start_pc, regs, exec_mask, check_range=check_range,
-                        memory_bounds=memory_bounds, max_steps=max_steps)
-
-  def _blocks(self, program: tuple[IR3Instruction, ...]) -> dict[int, int]:
-    # decode_ir3 caches and reuses each exact program tuple.  Key by identity
-    # here so the hot scheduler never hashes thousands of decoded instructions.
-    program_id = id(program)
-    if (cached := self._get_lru(self.program_blocks, program_id)) is not None and cached[0] is program: return cached[1]
-    blocks: dict[int, int] = {}
-    pc = 0
-    while pc < len(program):
-      if not self._supported(program[pc]):
-        pc += 1
-        continue
-      start = pc
-      while pc < len(program) and self._supported(program[pc]): pc += 1
-      end = pc
-      for candidate in range(start, end):
-        stop, count = candidate, 0
-        while stop < end and count < self.max_instructions:
-          count += program[stop].name != 'nop'
-          stop += 1
-        if count >= self.min_instructions: blocks[candidate] = stop
-    if (evicted := self._put_lru(self.program_blocks, program_id, (program, blocks), self.max_programs)) is not None:
-      self.stats.cache_evictions += 1
-      evicted_id = evicted[0]
-      for policy_key in tuple(self.program_policy):
-        if policy_key[0] == evicted_id: del self.program_policy[policy_key]
-      for key in tuple(self.cache):
-        if key[0] == evicted_id: del self.cache[key]
-      for key in tuple(self.uncompilable):
-        if key[0] == evicted_id: del self.uncompilable[key]
-    return blocks
-
-  def can_run_blocks(self, program: tuple[IR3Instruction, ...], lanes: int = 1) -> bool:
-    """Reject programs whose native chunks spend more time marshalling registers than interpreting IR3."""
-    program_id, wide = id(program), lanes >= 16
-    policy_key = (program_id, wide)
-    if (cached := self._get_lru(self.program_policy, policy_key)) is not None and cached[0] is program: return cached[1]
-    blocks = self._blocks(program)
-    ranges, covered_until = [], 0
-    for start, end in sorted(blocks.items()):
-      if start < covered_until: continue
-      ranges.append((start, end))
-      covered_until = end
-    pressures = [len(_Lowerer(program[start:end], 1).keys) for start, end in ranges]
-    average_limit = self.max_register_slots if wide else self.max_average_register_slots
-    enabled = bool(pressures) and sum(pressures) <= average_limit * len(pressures) and \
-      max(pressures) <= self.max_register_slots
-    self._put_lru(self.program_policy, policy_key, (program, enabled), self.max_programs * 2)
-    return enabled
-
-  def try_run(self, program: tuple[IR3Instruction, ...], start_pc: int, regs: dict[Register, list[int]],
-              exec_mask: list[bool], predication: list[bool] | None = None,
-              mask_pcs: frozenset[int] | None = None, *, policy_checked: bool = False) -> int | None:
-    """Return the next PC when accelerated, or ``None`` for exact interpreter fallback."""
-    if not regs: return None
-    lanes = len(next(iter(regs.values())))
-    if not policy_checked and not self.can_run_blocks(program, lanes): return None
-    self.stats.block_attempts += 1
-    if (end_pc := self._blocks(program).get(start_pc)) is None: return None
-    # A pending branch reconvergence inside the range would change the write mask mid-block;
-    # leave those ranges to the interpreter, which re-enters native execution after the merge.
-    if mask_pcs is not None and any(start_pc < target < end_pc for target in mask_pcs): return None
-    if len(exec_mask) != lanes: return None
-    mask = exec_mask if predication is None else [active and pred for active, pred in zip(exec_mask, predication, strict=True)]
-    key = (id(program), start_pc, end_pc, lanes)
-    try:
-      if (block := self._get_lru(self.cache, key)) is None and key not in self.uncompilable:
-        lowerer = _Lowerer(program[start_pc:end_pc], lanes)
-        if len(lowerer.keys) > self.max_register_slots: raise UnsupportedIR3Block('native IR3 block register pressure is too high')
-        signature = (lanes, lowerer.keys, tuple((inst.name, inst.dst, inst.srcs, inst.repeat, inst.repeat_srcs,
-                                                inst.src_mods, inst.condition, inst.types, inst.sat, inst.rounding,
-                                                inst.source_half) for inst in lowerer.instructions))
-        block = self._get_lru(self.compiled, signature)
-        if block is None:
-          if lanes >= 16:
-            block_class = 'wide'
-            class_limit = self.max_compiled_blocks - self.max_narrow_compiled_blocks
-          else:
-            block_class = 'priority_narrow' if len(program) >= 256 else 'regular_narrow'
-            class_limit = self.max_regular_narrow_compiled_blocks if block_class == 'regular_narrow' else \
-              self.max_narrow_compiled_blocks - self.max_regular_narrow_compiled_blocks
-          class_count = sum(value == block_class for value in self.compiled_classes.values())
-          if class_count >= class_limit:
-            if block_class == 'regular_narrow' or not self._evict_compiled_class(block_class):
-              raise UnsupportedIR3Block(f'{block_class} native IR3 block compile budget exhausted')
-          if len(self.compiled) >= self.max_compiled_blocks and not self._evict_compiled_class(block_class):
-            raise UnsupportedIR3Block('native IR3 block compile budget exhausted')
-          block = lowerer.compile()
-          self.stats.block_compiles += 1
-          self.compiled[signature] = block
-          self.compiled_classes[signature] = block_class
-        if self._put_lru(self.cache, key, block, self.max_block_locations) is not None:
-          self.stats.cache_evictions += 1
-      if block is None: return None
-      block.run(regs, mask)
-      self.stats.block_runs += 1
-      return start_pc + block.length
-    except UnsupportedIR3Block:
-      self.stats.block_declines += 1
-      self._put_lru(self.uncompilable, key, None, self.max_block_locations * 2)
-      return None
